+++ .storybook/main.jsx
@@ -1,14 +1,20 @@
 module.exports = {
-  stories: ['../src/**/*.stories.js'],
+  stories: ['../src/**/*.stories.jsx'],
   addons: [
-    '@storybook/preset-create-react-app',
     '@storybook/addon-actions',
     '@storybook/addon-links',
     '@storybook/addon-storysource',
     '@storybook/addon-viewport/register',
     '@storybook/addon-a11y/register',
     'storybook-mobile',
-    'storybook-dark-mode/register',
+    'storybook-dark-mode',
     '@storybook/addon-docs'
-  ]
+  ],
+  framework: {
+    name: '@storybook/react-vite',
+    options: {}
+  },
+  docs: {
+    autodocs: true
+  }
 }
